Move auth package under registry package
