+++ registry/auth/auth.go
