+++ registry/auth/token/accesscontroller.go
@@ -11,7 +11,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/docker/distribution/auth"
+	"github.com/docker/distribution/registry/auth"
 	ctxu "github.com/docker/distribution/context"
 	"github.com/docker/libtrust"
 	"golang.org/x/net/context"
