+++ registry/auth/token/token.go
@@ -13,7 +13,7 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/libtrust"
 
-	"github.com/docker/distribution/auth"
+	"github.com/docker/distribution/registry/auth"
 )
 
 const (
